requirements.txt
aiohappyeyeballs         2.4.6
aiohttp                  3.11.13
aiohttp-retry            2.9.1
aiosignal                1.3.2
alembic                  1.14.1
APScheduler              3.11.0
attrs                    25.1.0
audioop-lts              0.2.1
bidict                   0.23.1
blinker                  1.9.0
cachetools               5.5.2
certifi                  2025.1.31
charset-normalizer       3.4.1
click                    8.1.8
colorzero                2.0
Flask                    3.1.0
Flask-Migrate            4.1.0
Flask-SocketIO           5.5.1
Flask-SQLAlchemy         3.1.1
frozenlist               1.5.0
google-api-core          2.24.2
google-api-python-client 2.164.0
google-auth              2.38.0
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.1
googleapis-common-protos 1.69.2
gpiozero                 2.0.1
greenlet                 3.1.1
gTTS                     2.5.4
h11                      0.14.0
httplib2                 0.22.0
idna                     3.10
itsdangerous             2.2.0
Jinja2                   3.1.5
joblib                   1.4.2
Mako                     1.3.9
MarkupSafe               3.0.2
MouseInfo                0.1.3
multidict                6.1.0
numpy                    2.2.3
oauthlib                 3.2.2
paho-mqtt                2.1.0
pandas                   2.2.3
pillow                   11.1.0
pip                      24.2
propcache                0.3.0
proto-plus               1.26.1
protobuf                 6.30.1
pyasn1                   0.6.1
pyasn1_modules           0.4.1
PyAudio                  0.2.14
PyAutoGUI                0.9.54
PyGetWindow              0.0.9
PyJWT                    2.10.1
PyMsgBox                 1.0.9
pyparsing                3.2.1
pyperclip                1.9.0
PyRect                   0.2.0
PyScreeze                1.0.1
python-dateutil          2.9.0.post0
python-engineio          4.11.2
python-socketio          5.12.1
python3-xlib             0.15
pyttsx3                  2.98
pytweening               1.2.0
pytz                     2025.1
requests                 2.32.3
requests-oauthlib        2.0.0
RPi.GPIO                 0.7.1
rsa                      4.9
scikit-learn             1.6.1
scipy                    1.15.2
setuptools               75.8.0
simple-websocket         1.1.0
six                      1.17.0
SpeechRecognition        3.14.1
SQLAlchemy               2.0.38
standard-aifc            3.13.0
standard-chunk           3.13.0
threadpoolctl            3.5.0
twilio                   9.4.6
typing_extensions        4.12.2
tzdata                   2025.1
tzlocal                  5.3
uritemplate              4.1.1
urllib3                  2.3.0
Werkzeug                 3.1.3
wsproto                  1.2.0
yarl                     1.18.3
yt-dlp                   2025.2.19

